Prepare for the merge
