+++ expo-maps/android/src/main/java/expo/modules/maps/googleMaps/GoogleMapsMarkers.kt
@@ -1,4 +1,4 @@
-package expo.modules.maps
+package expo.modules.maps.googleMaps
 
 import android.net.Uri
 import com.google.android.gms.maps.GoogleMap
@@ -6,6 +6,8 @@ import com.google.android.gms.maps.model.BitmapDescriptorFactory
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.Marker
 import com.google.android.gms.maps.model.MarkerOptions
+import expo.modules.maps.MarkerObject
+import expo.modules.maps.interfaces.Markers
 
 class GoogleMapsMarkers(map: GoogleMap) : Markers {
