+++ expo-maps/android/src/main/java/expo/modules/maps/googleMaps/GoogleMapsPolylines.kt
@@ -1,7 +1,11 @@
-package expo.modules.maps
+package expo.modules.maps.googleMaps
 
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.model.*
+import expo.modules.maps.*
+import expo.modules.maps.Cap
+import expo.modules.maps.PatternItem
+import expo.modules.maps.interfaces.Polylines
 
 class GoogleMapsPolylines(map: GoogleMap) : Polylines {
   private val polylines = mutableListOf<Polyline>()
