+++ expo-maps/android/src/main/java/expo/modules/maps/interfaces/Circles.kt
@@ -1,4 +1,6 @@
-package expo.modules.maps
+package expo.modules.maps.interfaces
+
+import expo.modules.maps.CircleObject
 
 interface Circles {
   fun setCircles(circleObjects: Array<CircleObject>)
