+++ expo-maps/android/src/main/java/expo/modules/maps/interfaces/Polygons.kt
@@ -1,4 +1,6 @@
-package expo.modules.maps
+package expo.modules.maps.interfaces
+
+import expo.modules.maps.PolygonObject
 
 interface Polygons {
   fun setPolygons(polygonObjects: Array<PolygonObject>)
