+++ expo-maps/android/src/main/java/expo/modules/maps/interfaces/Polylines.kt
@@ -1,4 +1,6 @@
-package expo.modules.maps
+package expo.modules.maps.interfaces
+
+import expo.modules.maps.PolylineObject
 
 interface Polylines {
   fun setPolylines(polylineObjects: Array<PolylineObject>)
