+++ expo-maps/ios/ExpoMaps/AppleMaps/CustomTypes/ExpoAppleMapsPolyline.swift
