+++ expo-maps/ios/ExpoMaps/Protocols/Gestures.swift
